chore: fix additional lint errors
