+++ components/layout/DashboardLayout.tsx
@@ -1,7 +1,9 @@
-
 import { cookies } from 'next/headers';
 import Header from './Header';
 import Sidebar from './Sidebar';
+import EnvironmentBanner from './EnvironmentBanner';
+import CommandPaletteProvider from './CommandPaletteProvider';
+import PrintAuditReport from '../features/transactions/PrintAuditReport';
 import { SESSION_COOKIE_NAME, verifySessionToken } from '@/lib/auth/session';
 import type { UserRole } from '@/types';
 
@@ -13,53 +15,8 @@ async function getCurrentRole(): Promise<UserRole> {
   return session?.role ?? 'operator';
 }
 
-async function DashboardLayout({ children }: { children: React.ReactNode }) {
+export default async function DashboardLayout({ children }: { children: React.ReactNode }) {
   const role = await getCurrentRole();
-
-  return (
-    <div className="flex h-screen bg-gray-100">
-      <a
-        href="#main-content"
-        className="sr-only focus:not-sr-only focus:absolute focus:top-4 focus:left-4 focus:z-50 focus:px-4 focus:py-2 focus:bg-indigo-600 focus:text-white focus:rounded-md focus:shadow-lg"
-      >
-        Skip to main content
-      </a>
-      <Sidebar role={role} />
-      <div className="flex-1 flex flex-col overflow-hidden">
-        <Header role={role} />
-        <main
-          id="main-content"
-          className="flex-1 overflow-x-hidden overflow-y-auto bg-gray-100 p-6"
-          tabIndex={-1}
-
-"use client";
-
-import { useEffect, useState } from "react";
-import Header from "./Header";
-import Sidebar from "./Sidebar";
-import EnvironmentBanner from "./EnvironmentBanner";
-import CommandPalette from "./CommandPalette";
-import PrintAuditReport from "../features/transactions/PrintAuditReport";
-
-function DashboardLayout({ children }: { children: React.ReactNode }) {
-  const [isPaletteOpen, setIsPaletteOpen] = useState(false);
-
-  useEffect(() => {
-    const handleOpen = () => setIsPaletteOpen(true);
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if ((e.metaKey || e.ctrlKey) && e.key?.toLowerCase() === "k") {
-        e.preventDefault();
-        setIsPaletteOpen((prev) => !prev);
-      }
-    };
-
-    window.addEventListener("open-command-palette", handleOpen);
-    window.addEventListener("keydown", handleKeyDown);
-    return () => {
-      window.removeEventListener("open-command-palette", handleOpen);
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, []);
 
   return (
     <>
@@ -67,14 +24,13 @@ function DashboardLayout({ children }: { children: React.ReactNode }) {
         <a
           href="#main-content"
           className="sr-only focus:not-sr-only focus:absolute focus:top-4 focus:left-4 focus:z-50 focus:px-4 focus:py-2 focus:bg-indigo-600 focus:text-white focus:rounded-md focus:shadow-lg"
-
         >
           Skip to main content
         </a>
-        <Sidebar />
+        <Sidebar role={role} />
         <div className="flex-1 flex flex-col overflow-hidden">
           <EnvironmentBanner />
-          <Header />
+          <Header role={role} />
           <main
             id="main-content"
             className="flex-1 overflow-x-hidden overflow-y-auto bg-gray-100 p-6"
@@ -85,11 +41,8 @@ function DashboardLayout({ children }: { children: React.ReactNode }) {
         </div>
       </div>
 
-      <CommandPalette isOpen={isPaletteOpen} onClose={() => setIsPaletteOpen(false)} />
+      <CommandPaletteProvider />
       <PrintAuditReport />
     </>
   );
 }
-
-export default DashboardLayout;
-
